Improve 4.5.2 build

+++ scripts/cibuild/ConfigureVstsBuild.ps1
@@ -34,7 +34,9 @@ param
     [string]$FunctionalTestBuildId,
 
     [Parameter(Mandatory=$True)]
-    [string]$VstsRestApiRootUrl
+    [string]$VstsRestApiRootUrl,
+
+    [switch]$SkipUpdateBuildNumber
 )
 
 Function Get-Version {
@@ -131,41 +133,55 @@ if ($BuildRTM -eq 'true')
     # Set the $(NupkgOutputDir) build variable in VSTS build
     Write-Host "##vso[task.setvariable variable=NupkgOutputDir;]ReleaseNupkgs"
     Write-Host "##vso[task.setvariable variable=VsixPublishDir;]VS15-RTM"
-    $numberOfTries = 0
-    do{
-        Write-Host "Waiting for buildinfo.json to be generated..."
-        $numberOfTries++
-        Start-Sleep -s 15
-    }
-    until ((Test-Path $BuildInfoJsonFile) -or ($numberOfTries -gt 50))
-    $json = (Get-Content $BuildInfoJsonFile -Raw) | ConvertFrom-Json
-    $currentBuild = [System.Decimal]::Parse($json.BuildNumber)
-    # Set the $(Revision) build variable in VSTS build
-    Write-Host "##vso[task.setvariable variable=Revision;]$currentBuild"
-    Write-Host "##vso[build.updatebuildnumber]$currentBuild" 
-    $oldBuildOutputDirectory = Split-Path -Path $BuildInfoJsonFile
-    $branchDirectory = Split-Path -Path $oldBuildOutputDirectory
-    $newBuildOutputFolder =  Join-Path $branchDirectory $currentBuild
-    if(Test-Path $newBuildOutputFolder)
+    # Only for backward compatibility with orchestrated builds
+    if(-not $SkipUpdateBuildNumber)
     {
-        Move-Item -Path $BuildInfoJsonFile -Destination $newBuildOutputFolder
-        Remove-Item -Path $oldBuildOutputDirectory -Force
+        $numberOfTries = 0
+        do{
+            Write-Host "Waiting for buildinfo.json to be generated..."
+            $numberOfTries++
+            Start-Sleep -s 15
+        }
+        until ((Test-Path $BuildInfoJsonFile) -or ($numberOfTries -gt 50))
+        $json = (Get-Content $BuildInfoJsonFile -Raw) | ConvertFrom-Json
+        $currentBuild = $json.BuildNumber
+        # Set the $(Revision) build variable in VSTS build
+        Write-Host "##vso[task.setvariable variable=Revision;]$currentBuild"
+        Write-Host "##vso[build.updatebuildnumber]$currentBuild" 
+        $oldBuildOutputDirectory = Split-Path -Path $BuildInfoJsonFile
+        $branchDirectory = Split-Path -Path $oldBuildOutputDirectory
+        $newBuildOutputFolder =  Join-Path $branchDirectory $currentBuild
+        if(Test-Path $newBuildOutputFolder)
+        {
+            Move-Item -Path $BuildInfoJsonFile -Destination $newBuildOutputFolder
+            Remove-Item -Path $oldBuildOutputDirectory -Force
+        }
+        else
+        {
+            Rename-Item $oldBuildOutputDirectory $currentBuild
+        }
     }
-    else
-    {
-        Rename-Item $oldBuildOutputDirectory $currentBuild
-    }
-    
 }
 else
 {
-    $revision = Get-Content $BuildCounterFile
-    $newBuildCounter = [System.Decimal]::Parse($revision)
-    $newBuildCounter++
-    Set-Content $BuildCounterFile $newBuildCounter
-    # Set the $(Revision) build variable in VSTS build
-    Write-Host "##vso[task.setvariable variable=Revision;]$newBuildCounter"
-    Write-Host "##vso[build.updatebuildnumber]$newBuildCounter"
+    # Only for backward compatibility with orchestrated builds
+    if(-not $SkipUpdateBuildNumber)
+    {
+        $revision = Get-Content $BuildCounterFile
+        $newBuildCounter = [System.Decimal]::Parse($revision)
+        $newBuildCounter++
+        Set-Content $BuildCounterFile $newBuildCounter
+        # Set the $(Revision) build variable in VSTS build
+        $productVersion = $productVersion.Trim()
+        $FullBuildNumber = "$productVersion.$newBuildCounter"
+        Write-Host "##vso[task.setvariable variable=Revision;]$newBuildCounter"
+        Write-Host "##vso[build.updatebuildnumber]$FullBuildNumber"
+    }
+    else
+    {
+        $newBuildCounter = $env:BUILD_BUILDNUMBER
+    }
+
     $jsonRepresentation = @{
         BuildNumber = $newBuildCounter
         CommitHash = $env:BUILD_SOURCEVERSION
@@ -182,5 +198,5 @@ else
         Write-Error "Failed to get product version."
         exit 1
     }
-    Update-VsixVersion -manifestName source.extension.vs15.vsixmanifest -ReleaseProductVersion $productVersion -buildNumber $newBuildCounter
+    Update-VsixVersion -manifestName source.extension.vs15.vsixmanifest -ReleaseProductVersion $productVersion -buildNumber $env:BUILDNUMBER
 }
